wip: moved experiment to the marlowe-object-flow example
